+++ tables.sql
@@ -2,12 +2,12 @@ DROP TABLE IF EXISTS account CASCADE;
 
 CREATE TABLE account (
     account_id INT NOT NULL,
-    balance INT DEFAULT 0,
+    balance DOUBLE PRECISION DEFAULT 0,
     PRIMARY KEY (account_id)
 );
 
 CREATE TABLE position (
     symbol VARCHAR(256),
-    amount DOUBLE DEFAULT 0,
+    amount DOUBLE PRECISION DEFAULT 0,
     FOREIGN KEY (account_id) REFERENCES STATE(account_id) ON DELETE SET NULL ON UPDATE CASCADE
 );
